st004 Cleaning up the api code by introducing routes file
